uniswap v3 token prices

+++ src/Nethermind/Nethermind.Dsl.Test/UniswapSourceTests.cs
@@ -67,7 +67,7 @@ namespace Nethermind.Dsl.Test
             var denominator = scalarNumerator * inputNumerator;
 
             Fraction price = Fraction.FromDouble(numerator / denominator);
-            var x = price.ToDecimal().ToString();
+            var x = price.ToDouble().ToString();
         }
     }
 }
